Factor reflection out into its own file
No functional changes.
